grpc: move NameValueTemplatePair to common/schema
It will be needed there as well.
