+++ modules/grpc/common/schema/grpc-schema.hpp
@@ -27,8 +27,39 @@
 
 #include "syslog-ng.h"
 
+#include "compat/cpp-start.h"
+#include "template/templates.h"
+#include "compat/cpp-end.h"
+
 namespace syslogng {
 namespace grpc {
+
+struct NameValueTemplatePair
+{
+  std::string name;
+  LogTemplate *value;
+
+  NameValueTemplatePair(std::string name_, LogTemplate *value_)
+    : name(name_), value(log_template_ref(value_)) {}
+
+  NameValueTemplatePair(const NameValueTemplatePair &a)
+    : name(a.name), value(log_template_ref(a.value)) {}
+
+  NameValueTemplatePair &operator=(const NameValueTemplatePair &a)
+  {
+    name = a.name;
+    log_template_unref(value);
+    value = log_template_ref(a.value);
+
+    return *this;
+  }
+
+  ~NameValueTemplatePair()
+  {
+    log_template_unref(value);
+  }
+
+};
 
 class Schema
 {
